@@ -1,9 +1,9 @@
 class News {
-  String title;
-  String photo;
-  String description;
-  String date;
-  String author;
+  String? title;
+  String? photo;
+  String? description;
+  String? date;
+  String? author;
 
   News({this.title, this.photo, this.description, this.date, this.author});
 }
