@@ -1,10 +1,10 @@
 class VideoNews {
-  String title;
-  String video;
-  String thumbnail;
-  String description;
-  String date;
-  String author;
+  String? title;
+  String? video;
+  String? thumbnail;
+  String? description;
+  String? date;
+  String? author;
 
   VideoNews({this.title, this.thumbnail, this.video, this.description, this.date, this.author});
 }
